# docs/system_manager_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/status/services": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "List available services",
        "description": "Returns list of all services that can be monitored",
        "operationId": "list_services_status_services_get",
        "parameters": [
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicesListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/status/services/{service_id}/status": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "Get service status",
        "description": "Returns systemd status for a specific service",
        "operationId": "get_service_status_status_services__service_id__status_get",
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Service Id"
            }
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/status/services/{service_id}/logs": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "Get service logs",
        "description": "Returns the last N journal log lines for a specific service",
        "operationId": "get_service_logs_status_services__service_id__logs_get",
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Service Id"
            }
          },
          {
            "name": "lines",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Number of recent log lines to return",
              "default": 200,
              "title": "Lines"
            },
            "description": "Number of recent log lines to return"
          },
          {
            "name": "since_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1440,
              "minimum": 0,
              "description": "Only logs from last N minutes (0 = no filter)",
              "default": 0,
              "title": "Since Minutes"
            },
            "description": "Only logs from last N minutes (0 = no filter)"
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceLogsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/status/services/{service_id}/logs/stream": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "Stream service logs",
        "description": "Stream live journal logs for a specific service via journalctl --follow",
        "operationId": "stream_service_logs_status_services__service_id__logs_stream_get",
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Service Id"
            }
          },
          {
            "name": "since_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1440,
              "minimum": 0,
              "description": "Only logs from last N minutes (0 = no filter)",
              "default": 0,
              "title": "Since Minutes"
            },
            "description": "Only logs from last N minutes (0 = no filter)"
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/status/gpu/nvidia-smi": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "Get GPU status",
        "description": "Returns nvidia-smi output for GPUs",
        "operationId": "nvidia_smi_status_gpu_nvidia_smi_get",
        "parameters": [
          {
            "name": "detail",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Return detailed (-q) output",
              "default": false,
              "title": "Detail"
            },
            "description": "Return detailed (-q) output"
          },
          {
            "name": "gpu",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "GPU index or 'all'",
              "default": "all",
              "title": "Gpu"
            },
            "description": "GPU index or 'all'"
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NvidiaSmiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/status/gpu/reset": {
      "post": {
        "tags": [
          "status"
        ],
        "summary": "Reset NVIDIA GPU(s)",
        "description": "Resets GPU hardware/software state to clear CUDA state (e.g. after OOM). Requires miner/validator auth.",
        "operationId": "gpu_reset_status_gpu_reset_post",
        "parameters": [
          {
            "name": "gpu",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Comma-separated GPU indices or UUIDs (e.g. '0,1,2' or 'GPU-xxx'); use 'all' to reset all GPUs",
              "default": "all",
              "title": "Gpu"
            },
            "description": "Comma-separated GPU indices or UUIDs (e.g. '0,1,2' or 'GPU-xxx'); use 'all' to reset all GPUs"
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GpuResetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/status/overview": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "System overview",
        "description": "Returns combined status of all services and GPUs",
        "operationId": "overview_status_overview_get",
        "parameters": [
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/status/disk/space": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "Get directory sizes",
        "description": "Returns sizes of immediate subdirectories within a given path",
        "operationId": "get_disk_space_status_disk_space_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Directory path to analyze",
              "default": "/",
              "title": "Path"
            },
            "description": "Directory path to analyze"
          },
          {
            "name": "diagnostic",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Enable diagnostic mode for deep analysis",
              "default": false,
              "title": "Diagnostic"
            },
            "description": "Enable diagnostic mode for deep analysis"
          },
          {
            "name": "max_depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "description": "Maximum depth for diagnostic mode (default: 3)",
              "default": 3,
              "title": "Max Depth"
            },
            "description": "Maximum depth for diagnostic mode (default: 3)"
          },
          {
            "name": "top_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Show top N directories per level (default: 10)",
              "default": 10,
              "title": "Top N"
            },
            "description": "Show top N directories per level (default: 10)"
          },
          {
            "name": "cross_filesystems",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Cross filesystem boundaries (include mounted volumes)",
              "default": false,
              "title": "Cross Filesystems"
            },
            "description": "Cross filesystem boundaries (include mounted volumes)"
          },
          {
            "name": "X-Chutes-Hotkey",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Hotkey"
            }
          },
          {
            "name": "X-Chutes-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Nonce"
            }
          },
          {
            "name": "X-Chutes-Signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chutes-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiskSpaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServiceLogsResponse": {
        "properties": {
          "service": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Service",
            "description": "Service identifier and unit"
          },
          "requested_lines": {
            "type": "integer",
            "title": "Requested Lines",
            "description": "Number of log lines requested"
          },
          "returned_lines": {
            "type": "integer",
            "title": "Returned Lines",
            "description": "Number of log lines returned"
          },
          "stdout_truncated": {
            "type": "boolean",
            "title": "Stdout Truncated",
            "description": "Whether output was truncated"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Logs",
            "description": "Log entries"
          }
        },
        "type": "object",
        "required": [
          "service",
          "requested_lines",
          "returned_lines",
          "stdout_truncated",
          "logs"
        ],
        "title": "ServiceLogsResponse"
      },
...
      "ServiceStatusResponse": {
        "properties": {
          "service": {
            "$ref": "#/components/schemas/ServiceInfo"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceStatus"
              },
              {
                "type": "null"
              }
            ],
            "description": "Service status details"
          },
          "healthy": {
            "type": "boolean",
            "title": "Healthy",
            "description": "Whether service is healthy"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error details if status check failed"
          }
        },
        "type": "object",
        "required": [
          "service",
          "healthy"
        ],
        "title": "ServiceStatusResponse"
      },
      "ServicesListResponse": {
        "properties": {
          "services": {
            "items": {
              "$ref": "#/components/schemas/ServiceInfo"
            },
            "type": "array",
            "title": "Services",
            "description": "List of available services"
          }
        },
        "type": "object",
        "required": [
          "services"
        ],
        "title": "ServicesListResponse"
      },
...
    }
  }
}